Move checkpoint and restore commands to new files
